--- client/src/test/scala/view/baseconfiguration/TurnoTest.scala
@@ -1,24 +1,0 @@
-package view.baseconfiguration
-
-import javafx.scene.control.{Label, TitledPane}
-import junitparams.JUnitParamsRunner
-import org.junit.{After, Before, Test}
-import org.junit.runner.RunWith
-import view.driverviewoperations.TurnoDriver
-import view.launchview.DriverLaunch
-
-@RunWith(classOf[JUnitParamsRunner])
-class TurnoTest extends BaseTest {
-  var driverTurno:TurnoDriver=_
-  @Before
-  def beforeEachFerieTest(): Unit = {
-
-    setUp(classOf[DriverLaunch])
-    driverTurno = TurnoDriver(this)
-  }
-
-  @After
-  def closeScene():Unit={
-    closeCurrentWindow()
-  }
-}
